sun8i-ce: Enable H2/H3 TRNG as low quality entropy source
The H2/H3 TRNG peripheral is a highly biased entropy source, but its output
can still be used to seed the entropy pool. In my testing, the entropy
of the TRNG stream varied between 3-7 bits per byte. This patch enables
the H2/H3 TRNG with a low-ball estimate of 1 bit of entropy per byte.